requirements.txt
Click>=8.1.0,<9.0.0
colorama>=0.4.0,<1.0.0
pyautogen>=0.2.0,<0.3.0
